chore: Move `cmaes.py` from `core` to `baselines` (#200)
* Move cmaes.py from core to baselines
* Change corresponding api docs

---------

Authored-by: Lisa <lc1021@ic.ac.uk>
