+++ mkdocs.yml
@@ -154,11 +154,11 @@ nav:
         - DADS: api_documentation/core/dads.md
         - SAC: api_documentation/core/sac.md
         - TD3: api_documentation/core/td3.md
-        - CMAES: api_documentation/core/cmaes.md
         - Genetic Algorithm: api_documentation/core/genetic_algorithm.md
         - NSGA2: api_documentation/core/nsga2.md
         - SPEA2: api_documentation/core/spea2.md
         - PBT: api_documentation/core/pbt.md
+        - CMAES: api_documentation/core/cmaes.md
       - Containers: api_documentation/core/containers.md
       - Emitters: api_documentation/core/emitters.md
       - Neuroevolution: api_documentation/core/neuroevolution.md
