+++ qdax/core/emitters/cma_mega_emitter.py
@@ -6,7 +6,7 @@ from typing import Callable, Optional, Tuple
 import jax
 import jax.numpy as jnp
 
-from qdax.core.cmaes import CMAES, CMAESState
+from qdax.baselines.cmaes import CMAES, CMAESState
 from qdax.core.containers.mapelites_repertoire import (
     MapElitesRepertoire,
     get_cells_indices,
