+++ qdax/core/emitters/cma_rnd_emitter.py
@@ -6,7 +6,7 @@ from typing import Optional, Tuple
 import jax
 import jax.numpy as jnp
 
-from qdax.core.cmaes import CMAESState
+from qdax.baselines.cmaes import CMAESState
 from qdax.core.containers.mapelites_repertoire import MapElitesRepertoire
 from qdax.core.emitters.cma_emitter import CMAEmitter, CMAEmitterState
 from qdax.custom_types import Descriptor, ExtraScores, Fitness, Genotype, RNGKey
